third_party/WebKit/LayoutTests/fast/dom/HTMLImageElement/image-crossOrigin-expected.txt

This tests the handling of the img.crossOrigin attribute

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS 'crossOrigin' in img is true
PASS img.crossOrigin is null
PASS img.setAttribute('crossorigin', ''); img.crossOrigin is "anonymous"
PASS img.removeAttribute('crossorigin'); img.crossOrigin is null
PASS img.setAttribute('crossorigin', 'foo'); img.crossOrigin is "anonymous"
PASS img.setAttribute('crossorigin', 'ANONYMOUS'); img.crossOrigin is "anonymous"
PASS img.setAttribute('crossorigin', 'use-credentials'); img.crossOrigin is "use-credentials"
PASS img.crossOrigin = ''; img.getAttribute('crossorigin') is ""
PASS img.crossOrigin = 'foo'; img.getAttribute('crossorigin') is "foo"
PASS img.crossOrigin = null; img.hasAttribute('crossorigin') is false
PASS img.crossOrigin = 'ANONYMOUS'; img.getAttribute('crossorigin') is "ANONYMOUS"
PASS img.crossOrigin = 'use-credentials'; img.getAttribute('crossorigin') is "use-credentials"
PASS successfullyParsed is true

TEST COMPLETE
